Add files via upload
At the end of summer vacation, I felt I needed one more project to make it into the best summer ever! I found Robert Theaton's programming projects for advanced beginners and got right down to business. I had to play around a lot with the scaling and the sizing of the characters so I included how the finish version looks in case it differs from computer to computer. 
I just had to choose the magical Disney castle because at the end of the summer, who isn't thinking Disney!!!
https://robertheaton.com/2018/12/08/programming-projects-for-advanced-beginners/